Add Thread
We have a whiteboard test, but it's about designing the architecture for a real feature, no code. Our pair programming is inspired by a feature on a real site and is very representative of the sorts of work we do on a daily basis. Our take home test is less real-world, but designed to test code quality, not CS knowledge, and can be done in the candidates own time, on their machine, in their development environment.