Added Playlyfe
We don't ask brain-teasers. Since we mostly work in web technologies, we ask our candidates to solve simple problems in javascript (or if feasible, in their choice of language). We keep an eye one how the candidate codes. Most of the times, their capabilities and aptitude becomes very clear from the approach they take while solving the problem, or by their responses when they are stuck, or unable to solve the problem.