Update CRITERIA.md - NO UNPAID PROJECTS
Unpaid take home projects are wrong. We deserve a wage for labor that we provide, and the promise of a job someday isn't good enough, given the behavior of most hiring companies.

Unpaid take home projects make it hard for anyone who is already employed, has children, cares for a sick or elderly family member, etc. to be successful. Even when such a project is completed, the hiring party often ghosts candidates or doesn't provide feedback.  All of this benefits the company and not the potential employee.

Maybe some reasonable parameters around unpaid projects could be established (1hr? VM already set up w/ dev environment? Timeline for notification after the project is submitted? Feedback of some kind?)

But lacking that, we should disallow them altogether until there's some sanity in the process.

At the very least, we could add a label to all the companies in the list that require unpaid take home projects.