Remove Aurora Solar (#1187)
I interviewed at Aurora Solar a bout 3 months ago. They use a whiteboarding session. Seems like they're calling it "pair programming" but it's not that... the other person just asked a few questions and sat silently as I coded the solution.... so not pair programming. It's a white boarding session.