Add Poll Everywhere
Our technical assessment tries to mirror day-to-day contributions as much as we can. It asks the candidate to build a _very_ stripped down version of our core product experience. The follow-up interview is to assess team collaboration and pairs the candidate with a Designer and Engineer to enhance what they built in the challenge round.